make IM messages and chats persistent in the database.
They used to be in memory portion of the database, and are wiped when the device reboots or the android.acore process dies. We are trying to make them more resilient to process dying, as well as trying to retain longer message history.

updated database version, implemented upgrade path (creating Messages and Chats tables in persistent store).

added back the trigger to delete chat when a contact is deleted.