Removed packet_id unique constraint (issue 1221/2599)
The packet_id inside the ImProvider is marked as unique.
But it can happen often that this is not the case. This results
in messages not shown inside the Im application.

Because the messages are stored inside an in memory table, the
database version does not need to be upgraded.

More details at http://code.google.com/p/android/issues/detail?id=1221
or http://code.google.com/p/android/issues/detail?id=2599