run IM provider in gapps process instead of acore, so everything GTalkService needs to run as a persistent service lives inside a single process. This hopefully relieves some memory pressure in Donut.
The negative side effect of this is when the user upgrade to Donut, he will lose his old IM storage, which
in Donut contains any account username/pw for AIM/Yahoo/MSN, the provider settings (i.e. auto-login, notification
ringtones, etc), and the outgoing RMQ messages not yet received by MCS. The amount of data lost is not significant
and can all be recovered by the user. It's not he will lost chat messages because we don't store those in
persistent storage in Donut.

The things done to make IM provider running in gapps are:
- make IM provider use the google uid.
- make it run in the process "com.google.process.gapps"
- in the make file, make it use "vendor/google/certs/app" certificate.